+++ src/components/Theme/tests/Theme.test.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import Theme, { useTheme } from "components/Theme";
+import userEvent from "@testing-library/user-event";
 
 const Component = () => {
 	const { colorMode } = useTheme();
@@ -37,5 +38,65 @@ describe("Utilities/Theme", () => {
 		);
 
 		expect(screen.getByText("dark")).toBeInTheDocument();
+	});
+
+	test("changes parent theme", async () => {
+		const Component = () => {
+			const { theme, rootTheme, setTheme } = useTheme();
+
+			return (
+				<button data-theme={theme} data-root-theme={rootTheme} onClick={() => setTheme("foo")} />
+			);
+		};
+
+		render(
+			<Theme defaultName="reshaped" colorMode="inverted">
+				<Theme defaultName="slate" colorMode="inverted">
+					<Component />
+				</Theme>
+			</Theme>
+		);
+
+		const button = screen.getByRole("button");
+
+		expect(button).toHaveAttribute("data-theme", "slate");
+		expect(button).toHaveAttribute("data-root-theme", "reshaped");
+
+		await userEvent.click(button);
+
+		expect(button).toHaveAttribute("data-theme", "foo");
+		expect(button).toHaveAttribute("data-root-theme", "reshaped");
+	});
+
+	test("changes root theme", async () => {
+		const Component = () => {
+			const { theme, rootTheme, setRootTheme } = useTheme();
+
+			return (
+				<button
+					data-theme={theme}
+					data-root-theme={rootTheme}
+					onClick={() => setRootTheme("foo")}
+				/>
+			);
+		};
+
+		render(
+			<Theme defaultName="reshaped" colorMode="inverted">
+				<Theme defaultName="slate" colorMode="inverted">
+					<Component />
+				</Theme>
+			</Theme>
+		);
+
+		const button = screen.getByRole("button");
+
+		expect(button).toHaveAttribute("data-theme", "slate");
+		expect(button).toHaveAttribute("data-root-theme", "reshaped");
+
+		await userEvent.click(button);
+
+		expect(button).toHaveAttribute("data-theme", "slate");
+		expect(button).toHaveAttribute("data-root-theme", "foo");
 	});
 });