+++ src/components/Theme/useTheme.ts
@@ -10,17 +10,19 @@ export const useGlobalColorMode = () => {
 };
 
 export const useTheme = () => {
-	const { colorMode, theme, setTheme } = React.useContext(ThemeContext);
+	const { colorMode, theme, setTheme, rootTheme, setRootTheme } = React.useContext(ThemeContext);
 	const { mode, setMode, invertMode } = React.useContext(GlobalColorModeContext);
 
 	return React.useMemo(
 		() => ({
 			theme,
 			setTheme,
+			rootTheme,
+			setRootTheme,
 			colorMode: colorMode || mode,
 			setColorMode: setMode,
 			invertColorMode: invertMode,
 		}),
-		[colorMode, mode, setMode, invertMode, theme, setTheme]
+		[colorMode, mode, setMode, invertMode, theme, setTheme, setRootTheme, rootTheme]
 	);
 };