+++ src/utilities/dom.ts
@@ -3,7 +3,7 @@ export const getClosestFlyoutTarget = (el: HTMLElement | null): HTMLElement => {
 	const overflowY = style?.overflowY;
 	const position = style?.position;
 	const isScrollable = overflowY?.includes("scroll");
-	const isFixed = position === "fixed";
+	const isFixed = position === "fixed" || position === "sticky";
 
 	if (el === document.body || !el) return document.body;
 	if ((isScrollable && el.scrollHeight > el.clientHeight) || isFixed) return el;